Switched load on load
Switches will now display their appropriate state when the page first loads. However, the variables which determine if a switch is on or not resets on reload, so we need to either have those variables stored on the server, or check to see what state the switches should be in using a different method.